MB-35135: Deal with old PRE in trackedWrites due to partial snapshot
Consider the following scenario:
1) Replica is streaming from the active and receives a partial
   snapshot:
       [1:PRE(k1), 2:NOT RECEIVED(k2)]
   Importantly, not receiving the item at seqno 2 means that we do not
   move the HPS as we never received the snapshot end so this blocks
   us from removing 1:PRE at step 3a. It does not matter what sort of
   item we have at seqno 2.

2) Replica disconnects and reconnects which sets the
   allowedDuplicatePrepareSeqnos window to 1

3) Replica receives the following disk snapshot:
       [4:PRE(k1), 5:MUT(k1)]
   We have deduped the initial prepare and the commit at seqno 3.

   a) 4:PRE(k1)
      We replace 1:PRE in the HashTable with 4:PRE and add 4:PRE to
      trackedWrites in the PDM.
      This prepare logically completes 1:PRE in the PDM but 1:PRE is
      not removed from trackedWrites as the HPS used in the fence to
      remove the SyncWrites is still 0 and won't be moved until the
      snapshot end.
   b) 5:MUT(k1)
      We find 4:PRE in the HashTable and use this seqno when we attempt
      to complete the SyncWrite in the PDM. The PDM starts searching
      for the SyncWrite to complete at the beginning of trackedWrites
      as we are in a disk snapshot and must allow out of order
      completion. We then find the trackedWrite for 1:PRE that still
      exists in the PDM.

Due to out of order completion of prepares, we must start searching
for a SyncWrite to complete in trackedWrites from the beginning of
the list as opposed to the HighCompletedSeqno. If we have not
received a full snapshot before disconnect then we can have a
prepare in the PDM that has a seqno higher than the HPS as we only
move the HPS on snapshot boundary. If we then receive a disk
snapshot containing both a prepare and a commit for the same key
as the prepare before the disconnect then we will throw an exception
as we have two prepares in the PDM for the same key but only the
first is found which will have a different seqno to the one that we
are trying to complete.

Fix this by removing the original SyncWrite from the PDM when we add
an overwriting prepare.

Change-Id: Ie27b9073faea79c8b2737d7751521a67e66ad5c9
Reviewed-on: http://review.couchbase.org/112195
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>