MB-35190: Require ack from active before Majority SWs are satisfied
If Majority level SyncWrites can be satisfied without being acked by
the active (by being acked by enough replicas to reach majority) then
the ActiveDM may find SWs in the completedQueue which are satisfied,
but it has *not* acked.

For example,
  1                      2
PRE(PersistToMajority) PRE(Majority)

If persistence is slow, due to in-order committing, the ActiveDM
highPreparedSeqno may be waiting for the first PRE to persist.

If the timeout for both prepares expires, they would be moved to the
completed queue to be aborted.

BUT, the PRE(Majority) would be "satisfied". As a result,
processCompletedSyncWriteQueue would then try to commit with a seqno
greater than its highPreparedSeqno.

Requiring an ack from the active prevents this scenario, as the HPS is
advanced when the ActiveDM acks.

Change-Id: If6b81cfc15dc04efc4da188c2a938d3f1f1abc00
Reviewed-on: http://review.couchbase.org/112221
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>