MB-35208: Mark and ignore logically completed PDM SyncWrites
We keep prepares in the PDM trackedWrites list until they
have been persisted when they come from disk snapshots. This is
because they may have de-duped a PersistToMajority write. Even if
we commit this prepare in the disk snapshot we cannot remove it from
trackedWrites because the HPS will not be moved until we persist all
prepares in the disk snapshot. If we then receive another prepare for
the same key from an in-memory snapshot then we would throw an
exception as we have two SyncWrites for the same key in trackedWrites.

Mark prepares as completed if we receive a commit for them in a disk
snapshot and ignore any logically completed prepares in our sanity
check in PDM::addSyncWrite.

Change-Id: I969c15c356d51a3678c2e95cad40270d02803b5b
Reviewed-on: http://review.couchbase.org/112283
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>