MB-35208: Mark non-disk completed prepares as completed
Made a small oversight in my previous patch to mark and ignore
logically completed prepares and only made the code do this for
disk snapshots. We also need to do this for any PersistToMajority
prepares as well as the same race condition with persistence
exists. To make things simpler, we can mark all completed prepares
as completed.

Change-Id: I63a69007d8f7241797a52b4704a162d7c8a0c681
Reviewed-on: http://review.couchbase.org/112327
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>