MB-35189 [SR]: Transfer node positions on setReplicationTopology
When an ActiveDM has it's topology changed such that one (or more)
nodes from the old topology still exist in the new topology, the
trackedWrites iterators associated with each node in the two
replication chains are not updated. It is left as the default value of
list::end().

Given that list::end() for a node iterator indicates the iterator is
one-past-the-start, when that iterator is next checked to see what the
next SyncWrite needing acknowledgement is it will return
list::begin().

For example, consider the position of the active node (which normally
*is* present in both the old and new chains) for a trackedWrites of
2 SyncWrites, which has already acknowledged both SyncWrites:

    Before setReplicationTopology:

        | 1 | 2 |
              ^  ^
             pos |
                 |
                next
    After:

        | 1 | 2 |
          ^       ^
          |      pos
          |
         next

This is incorrect and can result in the ActiveDM attempting to
acknowedge the same SyncWrite twice. This manifests as the following
error when attempting to move highPreparedSeqno backwards:

    exception occurred in runloop during packet execution. ... Monotonic<x> (unlabelled) invariant failed: new value (533) breaks invariant on current value (534)

Fix by transferring over the iterator positions to the new chain, and
also ensure the ack count is transferred over.

Change-Id: I0804aa417d5deb4278e3e2fe7acf4b69b7930aca
Reviewed-on: http://review.couchbase.org/112423
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>