MB-35061: Ensure correct ConnHandler is removed
In ConnMap::removeVBConnByVBId_UNLOCKED we might break early if we find
a weak_ptr we could not lock.

If we do this, we might accidentally leave the connection we were
trying to remove (matched by cookie).

Methods like like DcpConnMap::seqnoAckVBPassiveStream assume there will
only be one ConnHandler found. If an "old" ConnHandler is left in the
ConnMap, these methods will find that first, and will perform their
action on the wrong ConnHandler.

Change-Id: Iba04e666f31e9d4dad33a217af1b096ca2ea13da
Reviewed-on: http://review.couchbase.org/112458
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>