MB-35245: Send items from Disk checkpoint as Disk snapshot
If a replica receives a Disk snapshot from an active then is
promoted it may then stream de-duped items from the
CheckpointManager. This is problematic as these items are sent by
ActiveStream as a Memory snapshot. This causes the replica to make
all sanity checks possible for durable writes that cannot be
satisfied if the original source of the items is disk. All of these
items are written into Disk checkpoints on the new active so we can
instead send Disk checkpoints as Disk snapshots to the replica which
can then make the appropriate sanity checks.

Currently, ActiveStream will attempt to fetch all items possible from
the CheckpointManager. Instead of doing this, modify CheckpointManager
to only return items from contiguous Checkpoints of the same type. In
the general case this will be functionally no different, but if a
newly promoted active has a Disk checkpoint then it will only fetch
items from Disk checkpoints. The next run of the
ActiveCheckpointProcessorTask will move the cursor into a Memory
checkpoint and fetch all items from Memory checkpoints. Also, when
becoming an active node, create a new checkpoint to ensure that we do
not write anything to a Disk checkpoint to avoid skipping any sanity
checks.

Change-Id: I853eda21f77972d81257e52a8766f486d0be808f
Reviewed-on: http://review.couchbase.org/112469
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>