MB-35275: Skip ADM::HPS update on null->topology change
When a replica is promoted to active, the trackedWrites are
moved from the PDM to the ADM. This ADM will have a null topology and
the active node iterator will not exist. When we move from a null
topology to a topology we would then update the HPS and this causes
the highPreparedSeqno to throw a monotonic exception as the HPS loops
from trackedWrites->end() back to trackedWrites->end() and commits
all prepares if possible.

Avoid calling updateHighPreparedSeqno post topology change and
instead iterate over trackedWrites and commit everything after
setting the HPS iterator correctly.

Change-Id: I8b74dfd307bebf2d3a596fdf9090275aa892ff61
Reviewed-on: http://review.couchbase.org/112550
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>