MB-35287: Handle responses to new producer messages
During rebalance/failover etc... the destination DCP
consumer stream can be closed whilst the producer is
still sending messages, the producer should expect
to receive responses to all new message types.

* Commit
* Prepare
* Abort
* SystemEvent

The commit improves logging in this area and classes
key-enoent as expected, however all other errors are
not expected and are logged as a warning.

Change-Id: I18ac613c8ad9afa088129f9b242bd0a39dbdbda1
Reviewed-on: http://review.couchbase.org/112551
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>