MB-35061: Remove DcpProducer from vbConns on sending StreamEnd
When `sendStreamEndOnClientStreamClose` is set, we do not immediately
remove a ConnHandler from vbConns, to allow us to shortly send a
StreamEnd. We erroneously *never* removed it in this case.

This became an issue when later removing a DcpConsumer - the old
DcpProducer would be found and removed instead. Then, with an "old"
DcpConsumer left in vbConns, we could add another DcpConsumer when
opening a stream. Several methods assume a single DcpConsumer per VB,
including `DcpConnMap::seqnoAckVBPassiveStream`.

This was first seen as an issue when a replica tried to seqno ack to
a node which was no longer the active for the vb.

The topology changes seen were (roughly) as follows for the three nodes
of interest.

Active Replica
A      B, C
B      A, C
C      A

Solution: Remove the DcpProducer once StreamEnd is sent.

Change-Id: Iabd52a5ca9a2d129159256529b388e03c87818e2
Reviewed-on: http://review.couchbase.org/112595
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>