MB-35314: Only negotiate SyncRepl if consumer has consumerName
ns_server will only create a consumer DCP connection with a
consumerName when the cluster is fully MadHatter+ and supports
SyncReplication. Currently the consumer code attempts to negotiate
SyncReplication with the producer node regardless of whether or not
it has a consumerName. It is only valid for the consumer node to
negotiate SyncReplication when it has a consumerName as the producer
needs to have received a replication topology. This is sent by
ns_server before we open DCP connections with consumerNames.

The current code causes rebalance failures due to lack of topology on
the active node as the consumer allows a seqnoAck to be sent as it
thinks that SyncReplication is enabled and so does the producer.

Change-Id: Ic8012c1e9fbd146351ea444c3c08c7975a505553
Reviewed-on: http://review.couchbase.org/112657
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>