MB-35367: Remove potential race in seqno acking
From code inspection, looks like we could enter either the
PDM::notifySnapshotEndReceived or PDM::notifyLocalPersistence
functions and acquire a hps value to ack back to the active.
The locking ensures that we never attempt to ack the same value
twice, but if the first thread were to reach the check if
hps != prevHps and get deschdeduled then a second thread could
make it through the lock and ack a high value before the first.
This would cause the active to throw monotonic invariant broken
exceptions.

Change-Id: Ib8cab59f8adb999302594f50057f327540e607c9
Reviewed-on: http://review.couchbase.org/112662
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>