MB-35061: Make addTakeoverStats tolerate missing stream
Calling DcpProducer::addTakeoverStats for a vb for which there is not a
stream should return does_not_exist. This is the case if the
StreamContainer is not present in the StreamMap, but if the
StreamContainer is present but empty an exception is thrown.

This was first seen causing rebalance to fail after an earlier patch
for MB-35061 ensured ConnHandlers are removed when intended.

Change-Id: I98f17e150543a493b51964f4d9f947b796fb9431
Reviewed-on: http://review.couchbase.org/112672
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>