MB-35332 [SR]: Set HPS::it correctly on ActiveDM -> PassiveDM conversion
An ActiveDurabilityMonitor can be converted to a
PassiveDurabilityMonitor during cluster topology changes - for example
during a rebalance where a node goes from active to replica. During
this conversion the highPreparedSeqno.iterator is not reset to the
correct position - it is left with the default value of
trackedWrites.end(). This is ok if trackedWrites is empty (as end() is
the correct value), but if trackedWrites is non-empty then it is
incorrct, and the next seqno to be considered for updating HPS will be
trackedWrites.begin():

    ActiveDM:

        | 1 | 2 |
              ^  ^
              |  |
              | next
             hps.iterator

    PassiveDM:

        | 1 | 2 |
          ^       ^
          |      hps.iterator
          |
         next

As a consequence, if the PassiveDM subsequently recieves another
prepare and attempts to update the Monotonic
highPreparedSeqno.lastWriteSeqno, it can essentially attempt to set it
to a value less than it's previous - 1 in the example above.

This manifests in the following error in the log:

    ERROR 71: exception occurred in runloop during packet execution. Cookie info: [...] - closing connection ([ 127.0.0.1:38045 - 127.0.0.1:11209 (<ud>@ns_server</ud>) ]):
    Monotonic<l> (unlabelled) invariant failed: new value (173) breaks invariant on current value (174)

Note this is similar to MB-35189 - where the ActiveDM didn't have the
node positions updated correctly on setReplicationTopology.

Fix is to ensure that the highPreparedSeqno (and highComittedSeqno)
iterators are correctly positioned when creating a PassiveDM from
existing trackedWrites.

Change-Id: Ic4d94316d49f029012d4906e76f10714e48c9f23
Reviewed-on: http://review.couchbase.org/112697
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>