MB-35226: Do not process seqno ack after stream is dead
We should not ack after a stream is dead because logically we
should not process any other messages. This previously could have
manifested in a where we receive a seqno ack for a replica that does
not exist in the replication topology and would have been added to
the queuedSeqnoAcks after a stream was dead. Setting the stream to
dead removes the queuedSeqnoAck so we would keep a queuedSeqnoAck
for a stream that no longer exists. If this replica were then to
reconnect and stream up to and ack the same sequence number, then
we would fire a monotonic invariant exception.

Change-Id: I976b4a1dedde58cde351ea543aca94e0f6370957
Reviewed-on: http://review.couchbase.org/112714
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>