MB-35361: Send seqno ack on stream (re)connection
If we have a vBucket with non-zero HPS accepting a passive stream
then we should send a SeqnoAcknowledgement to the active node to
ensure that we can commit any in-flight SyncWrites.

Consider the case where we have 1 replica and a durable write:

1) Replica receives prepare from active
2) Replica disconnects from active before acking
3) Replica attempts to ack but cannot due to lack of stream
4) Replica reconnects
5) Replica sends seqno ack

If we did not do step 5 then we may never complete the given prepare.
In the general case we would just timeout, but if this prepare had
been warmed up or grandfathered into an ADM then it would have
an infinite timeout and we would forever block this key if the
replica never received a disk snapshot or any other durable writes.

Change-Id: Ib54a0dda79745fdecfbce0a7517c6a468c25202a
Reviewed-on: http://review.couchbase.org/112781
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>