MB-35378: Ignore seqno ack on producer without stream
A stream on a producer may be removed from the streams map if
we have processed a close stream message and then stepped the
producer to send a StreamEnd. The consumer may not yet have
received and processed this StreamEnd but may have sent us a
seqno ack. If this is the case, we should simply ignore the
seqno ack.

Change-Id: I728e39993f6f556db61697bb29dd78338023f2ab
Reviewed-on: http://review.couchbase.org/112780
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>