Copy chain Positions between first and second chain
If we do not do this then the nodeWriteSeqno of a replica that
existed and acked in the secondChain then was moved to the first
will be incorrect until it acks again. This bug is benign as we use
the ack count to determine if we should commit, not the
nodeWriteSeqno so we do not fail to commit anything. No monotonic
invariant exceptions are thrown as a new Position is created in the
second chain which effectively resets the monotonic nodeWriteSeqno
value.

Change-Id: I881d4c66ed590169830656b1ca7979de37165967
Reviewed-on: http://review.couchbase.org/112844
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>