MB-35427: Remove ack for correct consumer
When we set the ActiveStream to dead we attempt to remove any
queuedSeqnoAck from the DM's map. We currently attempt to remove
the ack with the name of the ActiveStream which is not correct. We
should instead remove the ack with the name of the consumer which is
stored in the DcpProducer.

Change-Id: Ie4308b114542d61f2ffd91f6e65375f88114fe54
Reviewed-on: http://review.couchbase.org/112919
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>