MB-35429: Ack correct PassiveStream
The seqno ack code incorrectly assumes that we have only one
Consumer in the vbConns "map". This may not be the case if we have
a Consumer with dead PassiveStream and another Consumer with an
alive one.

Instead of attempting to ack for the first Consumer found, ack for
all Consumers found but only if the PassiveStream is alive.

Change-Id: If4adc3a6cb73ea3befaa23ad7bfd33bced0b7a65
Reviewed-on: http://review.couchbase.org/112961
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>