MB-35408: Add KVStore::prepareToCreate
Add a KVStore method that intends to improve on the changes made in
32119d47. KVStore::prepareToCreate is called when a new VBucket is being
created and allows the KVStore and the backend implementation to perform
specific tasks for the creation.

prepareToCreate now does the following:

* Resets the KVStore vbucket cache for the new VB.

For couch-kvstore backend

* increments the revision number for the vbucket

All other backends currently do no further work.

The only snag to this clean-up is the way KV-engine structures the
read-write and read-only KVStores, we have a cache per store, but the
couch-kvstore file revision structure is shared between a rw/ro pair.

The change ensures that prepareToCreate is invoked on both rw/ro stores,
but the couch-kvstore implementation knows to only increment the
revision if the store is rw.

Change-Id: Ibc578157f122b680ae00c4048730d3eda61d1e9f
Reviewed-on: http://review.couchbase.org/112965
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>