MB-35386: Update collections manifest. max_ttl is now maxTTL
The cluster REST API needs to expose the buckets collection
manifest and this should be the same logical JSON object that
KV is given via set-collections and exposed via get-collections.

However the cluster REST API requires max_ttl to be of the form
maxTTL, so let's update the collections JSON format to use maxTTL
so the same JSON can be used throughout.

Change-Id: Ib75271348a18417f2b8b9e7eb872f82f4fbb1fc0
Reviewed-on: http://review.couchbase.org/113010
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>