MB-35133: Cleanup old SyncWrite rollback code
Previously we attempted to rollback SyncWrites individually
by setting the correct state in the HashTable and PDM
post-rollback. With the new, simpler, approach to rollback where
we re-use the warmup code to load prepares we can remove the code
added to pass various state through to the PDM and the PDM rollback
function.

Change-Id: Ia605f04bbb5d32071a669564e58d4990809629ba
Reviewed-on: http://review.couchbase.org/113028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>