MB-35074: Fix mem_used tracking in get_collection_manifest
EventuallyPersistentEngine::get_collection_manifest calls a provided
response callback, which may grow a dynamic buffer as needed. The
buffer will later be free'd, and this deallocation will be tracked as
non-bucket.

Wrap the callback with makeExitBorderGuard to ensure the allocation is
not tracked against a bucket, never to be decremented.

Change-Id: Ibc4c8bd335bc7c362a84007acaf7c47d18263167
Reviewed-on: http://review.couchbase.org/113029
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>