Fix replica non-monotonic ack unit tests
The unit tests were written at the same time as a patch that
moved an exception from the Producer to ActiveStream but in a
different branch. Given we always hit the ActiveStream function,
we can now throw the exception in unit tests. Fix the tests to process
Items to ensure that we do not throw the exception.

Change-Id: I0f00e17da7a8ef2af2b4b1fa6e5dd83a02a2283a
Reviewed-on: http://review.couchbase.org/113097
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>