daemon/ExecutorPool: fix violations of ODR
The class ExecutorPool in daemon/executorpool.h has the same symbol
name as the one in engines/ep/src/executorpool.h. This violates the
One Definition Rule (ODR).

This problem has _apparently_ not been noticed thus far, however when
attempting to build with UBSan without optimization (-O0) it manifests
as an attempt to delete an invalid pointer.

Note the backtrace below - between frames #11 and #10 we flip between
the destructors of ep-ExecutorPool and daemon-ExecutorPool, this
results in attemptint to delete a pointer which is actually an object
vtable at frame #5:

    #5  0x00000000013c638f in std::_Destroy<...> (__pointer=0x7fffe5c276f0 <vtable for ExecutorPool+16>) at /usr/local/include/c++/7.3.0/bits/stl_construct.h:98
    ...
    #9  0x00000000013a043a in std::vector<...>::~vector (this=0x615000082080, __in_chrg=<optimized out>) at /usr/local/include/c++/7.3.0/bits/stl_vector.h:434
    #10 0x000000000139084c in ExecutorPool::~ExecutorPool (this=0x615000082080, __in_chrg=<optimized out>) at kv_engine/daemon/executorpool.h:35
    #11 0x00007fffe4425fe6 in ExecutorPool::~ExecutorPool (this=0x615000082080, __in_chrg=<optimized out>) at kv_engine/engines/ep/src/executorpool.cc:191
    #12 0x00007fffe44245df in ExecutorPool::shutdown () at kv_engine/engines/ep/src/executorpool.cc:151

Fix by moving the daemon ExecutorPool into the `cb` namespace (that
class arrived later and is used less, so is the easier one to rename).

Change-Id: Id6e5e154b9acc42d6b05e5cec5ec95ae72741462
Reviewed-on: http://review.couchbase.org/113139
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>