MB-34017: Send HCS in Disk snapshot
Send the HCS in Disk snapshots so that the replica node can
flush a correct HCS at the end of the snapshot which allows a
fast warmup (HCS to HPS) if we do not do any more SyncWrites
after our Disk snapshot.

In a future patch, we will pass this through to the flusher
via the CheckpointManager.

Change-Id: I36631b53611e903791643f7ce960960eea3f548f
Reviewed-on: http://review.couchbase.org/113187
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>