MB-35003: Enhance DCPLoopbackStreamTest for multiple replica tests
As part of the development of MB-35003 a diversion was taken into
validating how differing view of partial snapshots are  handled by
differing replicas. To fully validate some ideas a controlled multi
replica DCP test harness was required.

This commit updates the DCPLoopbackStreamTest so that the  tests can
create extra nodes and more easily manage the DCP producer/consumer
pairs.

In this change all of the DCP objects are held by a DcpRoute object
which expresses a route between two nodes. Multiple nodes can be created
(of which is fully demonstrated in a subsequent patch) and multiple
routes can be handled in a test.

Change-Id: I9b5677a3002d7df0e6af835a333649f5edf0211a
Reviewed-on: http://review.couchbase.org/113194
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>