MB-35003: Add a partial snapshots and failover test
Add a test that checks the following doesn't happen.

1. Two replicas receive different partial snapshots of the same data.
 a) Replica1 receives an in-memory stream
 b) Replica2 receives a de-duplicated disk stream
 c) Importantly Replica1 receives a key that Replica2 does not

2. Failover selects Replica1 to be active

Post failover, Replica2 is forced to rollback to before the  partial
snapshot and will then re-attempt to recieve the full snapshot.

Before the test the theory was that Replica2 may not rollback far enough
and that post failover Replica2 vs Replica3 would differ in the keys
they store. The test proved the theory was wrong, yet is a valuable test
to keep.

Change-Id: I5e4f7f14b2543367227135c59c77b2153c084612
Reviewed-on: http://review.couchbase.org/113195
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>