MB-35281: Port cbvbucketctl to use existing auth decorator
Brings cbvbucketctl in line with existing supported tools like cbstats,
avoiding unnecessarily duplicated code.

This will break existing usages of cbvbucketctl, but as it is an
unsupported tool this is a fair trade for consistent command line
tools, and avoiding the same bugs cropping up in several
reimplementations of the same thing.

Change-Id: Ic3075d76d5ab38ee8098ad5ffd30ec12466c231e
Reviewed-on: http://review.couchbase.org/113344
Tested-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>