MB-34955: Reduce the size of Monotonic instances when they do not throw
Move the std::string label to the throwing variant of 
Monotonic/AtomicMonotonic's base class (ThrowExceptionPolicy).

This results in only Monotonic instances which can throw pay the 
storage cost of the label. This means all instances when built 
CB_DEVELOPMENT_ASSERTS or when a Monotonic explicitly uses
the ThrowExceptionPolicy

The result: On a single node an empty bucket's (1024 vbuckets 
and CB_DEVELOPMENT_ASSERTS=false):

* mem_used is 60MB without this commit
* mem_used is 44MB with this commit

Change-Id: Ia0e02c13d8c59c9cc474de059c2ec2ef39614df1
Reviewed-on: http://review.couchbase.org/113372
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>