MB-35496: Remove completed prepares at compaction
When we compact a database file we can remove all completed
prepares to save space. A prepare is only needed for the lifetime
of a SyncWrite to ensure that we still have it should we shutdown
and restart. We don't need to worry about the metadata purge
interval when it comes to prepares because we stream Mutations
instead of Commits when streaming from disk.

Change-Id: I50925f11e72f3db38d92f5e0129d9a8dfe87b014
Reviewed-on: http://review.couchbase.org/113383
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>